# [7.x] Significant performance issue in Eloquent Collection loadCount() method