--- a/src/Illuminate/Database/Eloquent/Collection.php
+++ b/src/Illuminate/Database/Eloquent/Collection.php
@@ -89,6 +89,7 @@ class Collection extends BaseCollection implements QueueableCollection
 
         $this->each(function ($model) use ($models, $attributes) {
             $extraAttributes = Arr::only($models->get($model->getKey())->getAttributes(), $attributes);
+
             $model->forceFill($extraAttributes)->syncOriginalAttributes($attributes);
         });